fraud-detection-service/requirements.txt
kafka-python==2.0.2
redis==4.3.4
requests==2.28.1
flask==2.0.1
werkzeug==2.0.3 